+++ src/DallasTemperatureSensor.hpp
@@ -11,7 +11,7 @@
 
 class DallasTemperatureSensor : public Sensor {
 public:
-    DallasTemperatureSensor(int pin, std::size_t expectedNumberOfDevices);
+    DallasTemperatureSensor(uint8_t pin, std::size_t expectedNumberOfDevices);
 
     std::vector<std::string> measure() override;
 private:
